+++ core/test/Test/GHCJS.hs
@@ -6,11 +6,15 @@ import Prelude
 
 import Data.Foreign
 import Data.Maybe (isNothing)
+import Control.Concurrent (threadDelay)
+import Control.Monad.IO.Class (liftIO)
+import Data.IORef (newIORef, readIORef, writeIORef)
 import Halogen.VDom.DOM.Monad qualified as DOM
 import Test.Hspec (Spec, describe, it)
 import Test.Utils (assertEqual, assertWith)
-import Web.DOM.Internal.Types (Node (..))
+import Web.DOM.Internal.Types (Node (..), EventListener (..))
 import Web.DOM.ParentNode (ParentNode (..))
+import Web.Event.Event qualified as Event
 
 data DOMFixture
 
@@ -25,6 +29,8 @@ foreign import javascript unsafe "(() => { const child = {}; const sibling = { p
 foreign import javascript unsafe "((x) => x.child)" js_fixture_child_foreign :: Foreign DOMFixture -> Foreign Node
 foreign import javascript unsafe "((x) => x.sibling)" js_fixture_sibling_foreign :: Foreign DOMFixture -> Foreign Node
 foreign import javascript unsafe "((x) => x.parent)" js_fixture_parent_foreign :: Foreign DOMFixture -> Foreign ParentNode
+foreign import javascript unsafe "((cb) => { const t = new EventTarget(); t.addEventListener('submit', cb); return !t.dispatchEvent(new Event('submit', {cancelable:true})); })" js_cancelled :: EventListener -> IO Bool
+foreign import javascript unsafe "((cb) => { const t = new EventTarget(); let reached = false; t.addEventListener('click', cb); t.addEventListener('click', () => { reached = true; }); t.dispatchEvent(new Event('click')); return !reached; })" js_stopped :: EventListener -> IO Bool
 #else
 foreign import javascript unsafe "true" js_true :: Foreign Bool
 foreign import javascript unsafe "false" js_false :: Foreign Bool
@@ -36,6 +42,8 @@ foreign import javascript unsafe "const child = {}; const sibling = { previousSi
 foreign import javascript unsafe "$1.child" js_fixture_child_foreign :: Foreign DOMFixture -> Foreign Node
 foreign import javascript unsafe "$1.sibling" js_fixture_sibling_foreign :: Foreign DOMFixture -> Foreign Node
 foreign import javascript unsafe "$1.parent" js_fixture_parent_foreign :: Foreign DOMFixture -> Foreign ParentNode
+foreign import javascript unsafe "const t = new EventTarget(); t.addEventListener('submit', $1); return !t.dispatchEvent(new Event('submit', {cancelable:true}));" js_cancelled :: EventListener -> IO Bool
+foreign import javascript unsafe "const t = new EventTarget(); let reached = false; t.addEventListener('click', $1); t.addEventListener('click', () => { reached = true; }); t.dispatchEvent(new Event('click')); return !reached;" js_stopped :: EventListener -> IO Bool
 #endif
 
 js_fixture_child :: Foreign DOMFixture -> Node
@@ -49,6 +57,21 @@ js_fixture_parent = Node . js_fixture_parent_foreign
 
 spec :: Spec
 spec = describe "GHCJS FFI" $ do
+  it "cancels the default action before dispatch returns" $ do
+    listener <- DOM.runBrowserDOM $ DOM.mkEventListener Event.preventDefault
+    assertEqual "canceled" True =<< js_cancelled listener
+  it "stops later listeners during the same dispatch" $ do
+    listener <- DOM.runBrowserDOM $ DOM.mkEventListener Event.stopImmediatePropagation
+    assertEqual "stopped" True =<< js_stopped listener
+  it "continues a blocking event handler after canceling the event" $ do
+    finished <- newIORef False
+    listener <- DOM.runBrowserDOM $ DOM.mkEventListener $ \event -> do
+      Event.preventDefault event
+      liftIO $ threadDelay 10_000
+      liftIO $ writeIORef finished True
+    assertEqual "canceled before blocking" True =<< js_cancelled listener
+    threadDelay 50_000
+    assertEqual "handler resumed" True =<< readIORef finished
   it "converts true to True" $
     assertWith "foreignToBool should convert true to True" (foreignToBool js_true)
   it "converts false to False" $