+++ hooks/test/Test.hs
@@ -4,6 +4,7 @@ import Protolude
 import Test.Extra qualified as Extra
 import Test.Hooks qualified as Hooks
 import Test.Hspec (hspec)
+import Test.Scope qualified as Scope
 import Test.Storage qualified as Storage
 
 main :: IO ()
@@ -11,3 +12,4 @@ main = hspec $ do
   Hooks.spec
   Extra.spec
   Storage.spec
+  Scope.spec